remove external dependencies of core source altogether
