--- src/deps.ts
@@ -1,10 +1,0 @@
-// for src/m3/id-generation.ts:
-import {createHash} from "https://deno.land/std@0.177.0/node/crypto.ts"
-    // Note: this breaks from version 0.178.0 onwards!
-import {nanoid} from "npm:nanoid@4.0.2"
-
-export {
-    createHash,
-    nanoid
-}
-
